refactor: LottoTicket 테스트 코드 추가
6개 검증 및 중복 검증을 위한 단위테스트 코드 구현
