+++ README.md
@@ -12,7 +12,7 @@
 
 ### Domain
 
-* `Money`: 로또 구입 금액과 관련된 책임을 가진 객체이다. 로또 구입 금액을 활용하여 로또 티켓의 갯수를 계산할 수 있다.
+* `LottoPurchaseMoney`: 로또 구입 금액과 관련된 책임을 가진 객체이다. 로또 구입 금액을 활용하여 로또 티켓의 갯수를 계산할 수 있다.
     * [x] 구입 금액을 기반으로 티켓 갯수를 반환한다.
     * [x] 구입 금액이 1000원 미만인 경우 예외를 던진다.
     * [x] 거스름돈이 생긴 경우 가능한 티켓 갯수만 반환한다.
@@ -21,6 +21,9 @@
 * `LottoTicket`: 6자리의 로또 번호 관리의 책임을 가진 객체이다.
     * [x] 로또 티켓 생성 시점에 로또 번호 생성을 위한 전략을 활용한다.
     * [x] Getter로 꺼내오는 리스트를 변경할 경우 예외를 던진다.
+    * [x] LottoTicket을 기반으로 당첨 번호가 존재하는지 유무를 반환한다.
+    * [x] 로또 번호가 6개인지 검증한다.
+    * [x] 로또 번호 중복을 검증한다.
 
 * `LottoNumberGenerator`: 로또 번호를 생성 전략을 가진 인터페이스이다.
 
@@ -40,8 +43,6 @@
 
 * `WinningNumber`: 당첨 번호 한자리의 책임을 가지고 있다. 해당 당첨 번호는 보너스볼의 여부를 포함하고 있다.
     * [x] 해당 당첨 번호가 보너스볼인지 여부를 확인한다.
-    * [x] LottoTicket 을 기반으로 당첨 번호가 존재하는지 유무를 반환한다.
-    * [x] 1~45 사이의 숫자인지 검증한다.
 
 * `WinningNumbers`: 당첨 번호 리스트를 가진 일급 컬렉션이다.
     * [x] 당첨 번호 개수 만큼 당첨 번호를 생성한다.
