+++ src/test/java/lotto/domain/vo/LottoPurchaseMoneyTest.java
@@ -1,8 +1,7 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import static org.assertj.core.api.Assertions.*;
 
-import lotto.domain.vo.LottoPurchaseMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
